Benchmark multiple expressions in multiple environments

Measures the execution time for an R expression across multiple environments. The function runs the expression multiple times (controlled by `rep` parameter) and returns the execution times for each run in the specified environments.

Usage

benchInEnv(
  expr,
  envName = envList(),
  rep = 3,
  setup = NULL,
  resultAggregation = NULL
)

Arguments

expr

An R expression to be evaluated.

envName

A `character()` vector specifying the environment(s) name(s). If more than one environment is specified, the expression will be evaluated in each one.

rep

An `integer(1)` specifying the number of repetitions for each expression. Default is 3.

setup

An optional `expression` to be evaluated before the benchmarked expression. For instance, it can be used to load libraries or retrieve data.

resultAggregation

A function to aggregate the results. By default no aggregation is done, the results will be returned as a numeric vectors for each environment.

Value

A list of execution times for each environment:

A numeric vector of execution times for the evaluated expression if `length(envName) == 1`.

A list of numeric vectors of execution times for each environment if `length(envName) > 1`.

The result contains the user time (in seconds) for each repetition of the expression.

Note

This function is **experimental** and may change in future versions. Use at your own risk.

Examples

envCreate("my_env", packages = c("digest@0.6.36"))
#> - The project is out-of-sync -- use `renv::status()` for details.
#> # Downloading packages -------------------------------------------------------
#> - Querying repositories for available source packages ... Done!
#> - Downloading digest from Repository ...        OK [225.3 Kb in 0.37s]
#> Successfully downloaded 1 package in 1.9 seconds.
#> 
#> The following package(s) will be installed:
#> - digest [0.6.36]
#> These packages will be installed into "~/work/VerR/VerR/docs/reference/.envs/my_env/renv/library/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu".
#> 
#> # Installing packages --------------------------------------------------------
#> - Installing digest ...                         FAILED
#> Error : Error installing package 'digest':
#> ==================================
#> 
#> * installing *source* package ‘digest’ ...
#> ** this is package ‘digest’ version ‘0.6.36’
#> ** package ‘digest’ successfully unpacked and MD5 sums checked
#> ** using staged installation
#> ** libs
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> using C++ compiler: ‘g++ (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> g++ -std=gnu++17 -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2   -c SpookyV2.cpp -o SpookyV2.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c aes.c -o aes.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c blake3.c -o blake3.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c blake3_dispatch.c -o blake3_dispatch.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c blake3_portable.c -o blake3_portable.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c crc32.c -o crc32.o
#> g++ -std=gnu++17 -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2   -c crc32c.cpp -o crc32c.o
#> g++ -std=gnu++17 -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2   -c crc32c_portable.cpp -o crc32c_portable.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c digest.c -o digest.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c digest2int.c -o digest2int.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c init.c -o init.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c md5.c -o md5.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c pmurhash.c -o pmurhash.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c raes.c -o raes.o
#> raes.c: In function ‘AESFinalizer’:
#> raes.c:25:3: warning: implicit declaration of function ‘Free’; did you mean ‘free’? [-Wimplicit-function-declaration]
#>    25 |   Free(ctx);
#>       |   ^~~~
#>       |   free
#> raes.c: In function ‘AESinit’:
#> raes.c:42:23: warning: implicit declaration of function ‘Calloc’; did you mean ‘calloc’? [-Wimplicit-function-declaration]
#>    42 |   ctx = (aes_context*)Calloc(sizeof(*ctx), char);
#>       |                       ^~~~~~
#>       |                       calloc
#> raes.c:42:44: error: expected expression before ‘char’
#>    42 |   ctx = (aes_context*)Calloc(sizeof(*ctx), char);
#>       |                                            ^~~~
#> make: *** [/opt/R/4.5.0/lib/R/etc/Makeconf:202: raes.o] Error 1
#> ERROR: compilation failed for package ‘digest’
#> * removing ‘/home/runner/work/VerR/VerR/docs/reference/.envs/my_env/renv/staging/1/digest’
#> install of package 'digest' failed [error code 1]
#> Error: ! in callr subprocess.
#> Caused by error: 
#> ! Error installing package 'digest':
#> ==================================
#> 
#> * installing *source* package ‘digest’ ...
#> ** this is package ‘digest’ version ‘0.6.36’
#> ** package ‘digest’ successfully unpacked and MD5 sums checked
#> ** using staged installation
#> ** libs
#> using C compiler: ‘gcc (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> using C++ compiler: ‘g++ (Ubuntu 13.3.0-6ubuntu2~24.04) 13.3.0’
#> g++ -std=gnu++17 -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2   -c SpookyV2.cpp -o SpookyV2.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c aes.c -o aes.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c blake3.c -o blake3.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c blake3_dispatch.c -o blake3_dispatch.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c blake3_portable.c -o blake3_portable.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c crc32.c -o crc32.o
#> g++ -std=gnu++17 -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2   -c crc32c.cpp -o crc32c.o
#> g++ -std=gnu++17 -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2   -c crc32c_portable.cpp -o crc32c_portable.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c digest.c -o digest.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c digest2int.c -o digest2int.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c init.c -o init.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c md5.c -o md5.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c pmurhash.c -o pmurhash.o
#> gcc -std=gnu2x -I"/opt/R/4.5.0/lib/R/include" -DNDEBUG -I.  -I/usr/local/include    -fpic  -g -O2  -c raes.c -o raes.o
#> raes.c: In function ‘AESFinalizer’:
#> raes.c:25:3: warning: implicit declaration of function ‘Free’; did you mean ‘free’? [-Wimplicit-function-declaration]
#>    25 |   Free(ctx);
#>       |   ^~~~
#>       |   free
#> raes.c: In function ‘AESinit’:
#> raes.c:42:23: warning: implicit declaration of function ‘Calloc’; did you mean ‘calloc’? [-Wimplicit-function-declaration]
#>    42 |   ctx = (aes_context*)Calloc(sizeof(*ctx), char);
#>       |                       ^~~~~~
#>       |                       calloc
#> raes.c:42:44: error: expected expression before ‘char’
#>    42 |   ctx = (aes_context*)Calloc(sizeof(*ctx), char);
#>       |                                            ^~~~
#> make: *** [/opt/R/4.5.0/lib/R/etc/Makeconf:202: raes.o] Error 1
#> ERROR: compilation failed for package ‘digest’
#> * removing ‘/home/runner/work/VerR/VerR/docs/reference/.envs/my_env/renv/staging/1/digest’
#> install of package 'digest' failed [error code 1]
benchInEnv(Sys.sleep(1), "my_env", rep = 3, resultAggregation = mean)
#> [1] 1.001 1.001 1.001