Install new package(s) to environment(s)

Install new package(s) to environment(s)

Usage

envInstallPackage(package, envName = envList(), quiet = FALSE)

Arguments

package

A `character()` vector of package name(s) to install in the environment.

To specify the packages you want to install within the environment, you can use different syntax:

pkg: install the latest version of the package from CRAN.

pkg@version: install a specific version of the package from CRAN.

username/repo: install the package from GitHub. You can also specify the commit with username/repo@commitId.

bioc::pkg: install the package from Bioconductor.

Note that to install a specific version of a Bioconductor package, it is recommended to install it via GitHub.

You can find more information on package installation in the renv::install documentation.

envName

A `character()` string specifying the name(s) of the environment(s) for which to update the lockFiles. Default is all environments.

quiet

A `logical()` indicating whether messages should be suppressed. Default is `FALSE`.