Environment Management Functions

These functions allow users to create, clear, and list environments stored in the ".envs" directory.

Usage

createEnv(envName, packages = NULL, lockfile = NULL)

clearEnv(envName)

clearEnvs()

listEnvs()

copyToEnvs(sourcePath, envNames = listEnvs(), targetPath = "")

removeFromEnvs(targetPath, envNames = listEnvs())

exportLockfile(envName, exportPath = .lockFileStorageFromEnv(envName))

Arguments

envName

A `character(1)` string specifying the name of the environment.

packages

A `character()` vector of package names to install in the environment. Default is NULL.

lockfile

A `character()` string specifying the path to a lockfile. Default is NULL.

sourcePath

A `character(1)` string specifying the path of the file or directory to copy.

envNames

A `character()` that specify the environments' name in which the file/directory should be removed. Default is `listEnvs()` to remove the file/directory in all the environments.

targetPath

A `character(1)` string specifying the relative path within each environment where the file or directory should be removed.

exportPath

A `character(1)` string specifying the path where the lockfile should be exported. Default is "exportedLockFile/*envName*_lockFile.lock".

Value

`listEnvs()`: A `character()` vector that contains the environments' names

Details

Description of functions:

`createEnv()`: Creates an environment directory and installs the specified packages or uses the lockfile.

`clearEnv()`: Deletes a specific environment directory.

`clearEnvs()`: Deletes all environment directories.

`listEnvs()`: Returns a character vector containing the names of all existing environments.

Examples

createEnv("my_first_env", packages = c("digest@0.6.18"))
#> - The project is out-of-sync -- use `renv::status()` for details.
#> - The project is out-of-sync -- use `renv::status()` for details.
#> # Downloading packages -------------------------------------------------------
#> - Downloading rlang from CRAN ...               OK [1.5 Mb in 0.34s]
#> Successfully downloaded 1 package in 1.5 seconds.
#> 
#> The following package(s) will be installed:
#> - rlang [1.1.5]
#> These packages will be installed into "~/work/VerR/VerR/docs/reference/.envs/my_first_env/renv/library/linux-ubuntu-noble/R-4.4/x86_64-pc-linux-gnu".
#> 
#> # Installing packages --------------------------------------------------------
#> - Installing rlang ...                          OK [installed binary and cached in 0.28s]
#> Successfully installed 1 package in 0.34 seconds.
#> # Downloading packages -------------------------------------------------------
#> - Downloading digest from Repository ...        OK [125.5 Kb in 1.9s]
#> Successfully downloaded 1 package in 2.6 seconds.
#> 
#> The following package(s) will be installed:
#> - digest [0.6.18]
#> These packages will be installed into "~/work/VerR/VerR/docs/reference/.envs/my_first_env/renv/library/linux-ubuntu-noble/R-4.4/x86_64-pc-linux-gnu".
#> 
#> # Installing packages --------------------------------------------------------
#> - Installing digest ...                         OK [built from source and cached in 3.3s]
#> Successfully installed 1 package in 3.4 seconds.
#> - Lockfile written to "~/work/VerR/VerR/docs/reference/.envs/my_first_env/renv.lock".
#> The following package(s) will be updated in the lockfile:
#> 
#> # RSPM -----------------------------------------------------------------------
#> - digest   [0.6.18 -> *]
#> - renv     [* -> 1.1.2]
#> 
#> - Lockfile written to "~/work/VerR/VerR/docs/reference/.envs/my_first_env/renv.lock".
#> Environment created from package list in: /home/runner/work/VerR/VerR/docs/reference/.envs/my_first_env
#> NULL
createEnv("my_second_env", packages = c("digest@0.6.17"))
#> - The project is out-of-sync -- use `renv::status()` for details.
#> - The project is out-of-sync -- use `renv::status()` for details.
#> The following package(s) will be installed:
#> - rlang [1.1.5]
#> These packages will be installed into "~/work/VerR/VerR/docs/reference/.envs/my_second_env/renv/library/linux-ubuntu-noble/R-4.4/x86_64-pc-linux-gnu".
#> 
#> # Installing packages --------------------------------------------------------
#> - Installing rlang ...                          OK [linked from cache]
#> Successfully installed 1 package in 5.9 milliseconds.
#> # Downloading packages -------------------------------------------------------
#> - Downloading digest from Repository ...        OK [121.6 Kb in 1.9s]
#> Successfully downloaded 1 package in 2.6 seconds.
#> 
#> The following package(s) will be installed:
#> - digest [0.6.17]
#> These packages will be installed into "~/work/VerR/VerR/docs/reference/.envs/my_second_env/renv/library/linux-ubuntu-noble/R-4.4/x86_64-pc-linux-gnu".
#> 
#> # Installing packages --------------------------------------------------------
#> - Installing digest ...                         OK [built from source and cached in 3.0s]
#> Successfully installed 1 package in 3 seconds.
#> - Lockfile written to "~/work/VerR/VerR/docs/reference/.envs/my_second_env/renv.lock".
#> The following package(s) will be updated in the lockfile:
#> 
#> # RSPM -----------------------------------------------------------------------
#> - digest   [0.6.17 -> *]
#> - renv     [* -> 1.1.2]
#> 
#> - Lockfile written to "~/work/VerR/VerR/docs/reference/.envs/my_second_env/renv.lock".
#> Environment created from package list in: /home/runner/work/VerR/VerR/docs/reference/.envs/my_second_env
#> NULL
listEnvs()
#> [1] "my_first_env"  "my_second_env"
clearEnv("my_first_env")
#> Cleared environment: my_first_env
listEnvs()
#> [1] "my_second_env"
clearEnvs()
#> All environments cleared