Run expressions in environments

These functions allow execution of R expressions within specified environments.

`runInEnv(expr, envName)`: Runs an expression in a specific environment.

`runInEnvs(expr, envNames)`: Runs an expression across multiple environments.

Usage

runInEnv(expr, envName)

runInEnvs(expr, envNames = listEnvs())

Arguments

expr

An R expression to be evaluated.

envName

A `character(1)` string specifying the environment name.

envNames

A `character()` vector of environment names. Default is `listEnvs()`.

Value

The result of the evaluated expression for `runInEnv`.

A list of results for each environment for `runInEnvs`.

Examples

createEnv("my_env", packages = c("digest@0.6.18"))
#> - The project is out-of-sync -- use `renv::status()` for details.
#> # Downloading packages -------------------------------------------------------
#> - Querying repositories for available source packages ... Done!
#> - Downloading digest from Repository ...        OK [125.5 Kb in 1.6s]
#> Successfully downloaded 1 package in 2.8 seconds.
#> 
#> The following package(s) will be installed:
#> - digest [0.6.18]
#> These packages will be installed into "~/work/VerR/VerR/docs/reference/.envs/my_env/renv/library/linux-ubuntu-noble/R-4.4/x86_64-pc-linux-gnu".
#> 
#> # Installing packages --------------------------------------------------------
#> - Installing digest ...                         OK [built from source and cached in 3.2s]
#> Successfully installed 1 package in 3.3 seconds.
#> The following package(s) will be updated in the lockfile:
#> 
#> # CRAN -----------------------------------------------------------------------
#> - KernSmooth   [* -> 2.23-26]
#> - MASS         [* -> 7.3-64]
#> - Matrix       [* -> 1.7-2]
#> - boot         [* -> 1.3-31]
#> - class        [* -> 7.3-23]
#> - cluster      [* -> 2.1.8]
#> - codetools    [* -> 0.2-20]
#> - foreign      [* -> 0.8-88]
#> - lattice      [* -> 0.22-6]
#> - mgcv         [* -> 1.9-1]
#> - nlme         [* -> 3.1-167]
#> - nnet         [* -> 7.3-20]
#> - rpart        [* -> 4.1.24]
#> - spatial      [* -> 7.3-18]
#> - survival     [* -> 3.8-3]
#> 
#> # RSPM -----------------------------------------------------------------------
#> - digest       [* -> 0.6.18]
#> - renv         [* -> 1.1.2]
#> 
#> The version of R recorded in the lockfile will be updated:
#> - R            [* -> 4.4.3]
#> 
#> - Lockfile written to "~/work/VerR/VerR/docs/reference/.envs/my_env/renv.lock".
#> Environment created from package list in: /home/runner/work/VerR/VerR/docs/reference/.envs/my_env
#> NULL
runInEnv(packageVersion("digest"), "my_env")
#> - The project is out-of-sync -- use `renv::status()` for details.
#> [1] ‘0.6.18’
runInEnvs(packageVersion("digest"))
#> Running expression in environment: my_env 
#> - The project is out-of-sync -- use `renv::status()` for details.
#> $my_env
#> [1] ‘0.6.18’
#>