Benchmark expression in multiple environments

Measures the execution time for an R expression across multiple environments. The function runs the expression multiple times (controlled by `rep` parameter) and returns the execution times for each run in the specified environments.

Usage

benchInEnv(
  expr,
  envName = envList(),
  rep = 3,
  setup = NULL,
  resultAggregation = NULL
)

Arguments

expr

An R expression to be evaluated.

envName

A `character()` vector specifying the environment(s) name(s). If more than one environment is specified, the expression will be evaluated in each one.

rep

An `integer(1)` specifying the number of repetitions for each expression. Default is 3.

setup

An optional `expression` to be evaluated before the benchmarked expression. For instance, it can be used to load libraries or retrieve data.

resultAggregation

A function to aggregate the results. By default no aggregation is done, the results will be returned as a numeric vectors for each environment.

Value

A list of execution times for each environment:

A numeric vector of execution times for the evaluated expression if `length(envName) == 1`.

A list of numeric vectors of execution times for each environment if `length(envName) > 1`.

The result contains the user time (in seconds) for each repetition of the expression.

Examples

envCreate("my_env", packages = c("jsonlite"))
#> - The project is out-of-sync -- use `renv::status()` for details.
#> # Downloading packages -------------------------------------------------------
#> - Downloading jsonlite from https://packagemanager.posit.co/cran/__linux__/noble/latest ... OK [1 Mb in 0.12s]
#> Successfully downloaded 1 package in 0.84 seconds.
#> 
#> The following package(s) will be installed:
#> - jsonlite [2.0.0]
#> These packages will be installed into "~/work/VerR/VerR/docs/reference/.envs/my_env/renv/library/linux-ubuntu-noble/R-4.5/x86_64-pc-linux-gnu".
#> 
#> # Installing packages --------------------------------------------------------
#> - Installing jsonlite ...                       OK [installed binary and cached in 0.24s]
#> Successfully installed 1 package in 0.52 seconds.
#> Installed packages in environment: /home/runner/work/VerR/VerR/docs/reference/.envs/my_env
#> - The project is out-of-sync -- use `renv::status()` for details.
#> The following package(s) will be updated in the lockfile:
#> 
#> # RSPM -----------------------------------------------------------------------
#> - jsonlite   [* -> 2.0.0]
#> - renv       [* -> 1.1.4]
#> 
#> The version of R recorded in the lockfile will be updated:
#> - R          [* -> 4.5.0]
#> 
#> - Lockfile written to "~/work/VerR/VerR/docs/reference/.envs/my_env/renv.lock".
#> Updated lockfile in environment: .envs/my_env
benchInEnv(Sys.sleep(1), "my_env", rep = 3, resultAggregation = mean)
#> [1] 1.001 1.001 1.001